sm8150-common: ril-wrapper: implement workaround for RIL_UNSOL_LGE_SIGNAL_STRENGTH
LG has their own RIL stack (LgeRIL) in frameworks, as well as their own
HIDL (vendor.lge.hardware.radio).

Unfortunately, LG is also using that to send signal strength UNSOLs to
frameworks, so in our environment, the signal strength was never
received by the framework.

The reason this is relevant now is because, in Android 12 and lower,
AOSP had a fallback where it would manually poll the signal in the
event an UNSOL was never received; however, this was removed in Android
13.

Thankfully, much of LG's structs are largely reusing AOSP structs, so
we can in-place hot-swap the values as needed and start reporting the
correct UNSOL.

Change-Id: Id553ba52e4ed78f6b1c05007cb80666b4979af03